requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
astroid==2.6.6
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cohesion==1.0.0
colorama==0.4.4
configparser==5.0.2
cycler==0.11.0
dacite==1.6.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
Django==3.2
docker-pycreds==0.4.0
entrypoints==0.4
eradicate==2.0.0
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.5.0
flake8==3.9.0
flake8-broken-line==0.3.0
flake8-bugbear==21.4.3
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-eradicate==1.0.0
flake8-import-order==0.18.1
flake8-plugin-utils==1.3.1
flake8-polyfill==1.0.2
flake8-return==1.1.2
flake8-spellcheck==0.24.0
flake8-string-format==0.3.0
fonttools==4.29.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
huggingface-hub==0.0.8
hyperstyle==1.2.5
hyperstyle-analysis==1.2.1
idna==2.10
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
mando==0.6.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.2
openpyxl==3.0.7
packaging==21.3
pandarallel==1.5.2
pandas==1.2.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pep8-naming==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
plotly==4.14.3
pluggy==0.13.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.11.2
pylint==2.7.4
pylint-django==2.3.0
pylint-plugin-utils==0.7
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.3
pytest-runner==5.2
pytest-subtests==0.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
radon==4.5.0
regex==2022.1.18
requests==2.25.1
retrying==1.3.3
sacremoses==0.0.47
scikit-learn==0.24.2
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.5.5
serde==0.8.1
Serialize==0.1
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
subprocess32==3.5.4
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.10.2
toml==0.10.2
tomli==2.0.1
torch==1.8.1
tornado==6.1
tqdm==4.49.0
traitlets==5.1.1
transformers==4.6.1
typing-extensions==3.10.0.2
urllib3==1.26.8
wandb==0.10.31
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wps-light==0.15.2
wrapt==1.12.1
